remove WArning : comparison of integer expressions of different signedness, static_cast<uint>(rank()) , but perhaps rank() should return uint instead

+++ src/tensor_object.hpp
@@ -94,7 +94,7 @@ namespace jpacPhoto
 
         inline Type operator()(std::vector<lorentz_index> indices)
         {
-            if (indices.size() != rank()) return error("lorentz_tensor", "Incorrect number of indices passed!", NaN<Type>());
+	  if (indices.size() != static_cast<uint>(rank())) return error("lorentz_tensor", "Incorrect number of indices passed!", NaN<Type>());
             int mu = +indices[0];
             return _entries[mu];
         };
@@ -138,7 +138,7 @@ namespace jpacPhoto
 
         inline Type operator()(std::vector<lorentz_index> indices)
         {
-            if (indices.size() != rank()) return error("lorentz_tensor", "Incorrect number of indices passed!", NaN<Type>());
+	  if (indices.size() != static_cast<uint>(rank())) return error("lorentz_tensor", "Incorrect number of indices passed!", NaN<Type>());
             int mu = +indices[0], nu = +indices[1];
             return (mu == nu) ? complex((mu == 0) - (mu != 0)) * id : z;
         };
